// src/common/enums/motivosViajes.enum.ts
export enum MotivosViajes {
  NEGOCIOS_Y_MOTIVOS_PROFESIONALES = 'NEGOCIOS_Y_MOTIVOS_PROFESIONALES',
  VACACIONES_RECREO_Y_OCIO = 'VACACIONES_RECREO_Y_OCIO',
  VISITAS_A_FAMILIARES_Y_AMIGOS = 'VISITAS_A_FAMILIARES_Y_AMIGOS',
  EDUCACION_Y_FORMACION = 'EDUCACION_YFORMACION',
  SALUD_Y_ATENCION_MEDICA = 'SALUD_Y_ATENCION_MEDICA',
  RELIGION_Y_PEREGRINACIONES = 'RELIGION_Y_PEREGRINACIONES',
  COMPRAS = 'COMPRAS',
  TRANSITO = 'TRANSITO',
  OTROS_MOTIVOS = 'OTROS_MOTIVOS',
}